Cache userinfo and PIN.
This caches the user info from users/account in our plexprofile. It’s
obfuscated with AES.

We also cache the PIN for protected users in the same directory, but
this instead SHA256 hashed instead of AES obfuscated since we don’t
need to decode it later, we just need to make another hash to compare
them.