Revert "[vfs] [posix] SMBFile: fix reading files opened with READ_TRUNCATED flag"
This reverts commit b202cd09426d1ce1e8921b827a84476472cf35fe.

This change is invalid. The whole point of READ_TRUNCATED is to allow
lower layer file systems to return any amount of data less than or
equal to requested size. This is for efficiency reasons. After this
change each request of X bytes from xbmc would result in multiple
network roundtrips