[gui] update the view state's list reference before the view state is consulted about whether the "go to parent" item should be added to the list
Without this the view state will always think the previous directory is
the one you're currently in. This means that e.g. in the recordings
window the "go to parent" was shown in the root directory (except the
first time the window was entered since the path is set correctly then),
but not in subdirectories.
subdirectories.