Support recently unwatched items (optional)
This is a long requested feature for the Confluence skin that enables
recently unwatched items on the home screen for movies, tv shows and
albums instead of latest items. This does not add a dependency to
service.skin.widgets, but only uses it if its present. Works great.
Should also be ported to re-Touched.