Pivos: [build] Rework skin selection
This introduces an AC_ARG_WITH for skins that takes the form of a comma
separated list of skin addon IDs (eg. --with-skins=skin.skin1,skin.skin2,
...,skin.skinN) where the first skin declared in the list is forced to
be the default skin in app.

The benefits here should be obvious.  Changing the default skin in tree
is now a one liner in configure.in rather than several changes all over.
User/vendor builds can specify custom skins without modification.  Best
of all, the hardcoded hack job for touched skin can be removed as it is
entirely replicable by --with-skins=skin.touched.