Fix so that PHT can play more than 21 tracks.
At some point the code that refreshed the play queue window was removed
because it reloaded the PQ to often. This commit solves that by only
refreshing the PQ if we get a timeline with a low time option. That
means if it just started to play the song.

Fixes #1380