fixed: CID 719020 String not null terminated
The function may iterate past the end of the buffer looking for a null terminator.
In iso9660::​ReadRecursiveDirFromSector(unsigned int, char const*): A character buffer that has not been null terminated is passed to a function expecting a null terminated string (CWE-170)

Note, this also fixes an infinate loop in this parsing function.