Display user selection window when waking up from screensaver : part of #1414
This patch will show up StartupDialog when waking up from screensaver.

It will ony show the user selection and pin dialig if the current user uses a pin.

There is no setting to enable this yet or not, but a an options in screensaver prefs could be added, shoult it be required.