[pvr] fix: select the first item if we don't have a stored last selected item path
If you stay in the channel osd and select a different group, you never selected before, the selected item will be the same (group A = 22, group B = 22). This will fix the issue as it selects the first item instead of the same if we don't have a stored last selected item path.