[infoscanner] remove whitespaces from sample scan exclude regexs (fixes #15605)
This commit removes leading and trailing whitespaces from the movies and tvshows
exclude from scan regular expressions. Impact - if any - should be pretty low as
samples usually reside in a dedicated samples folder or are following naming
conventions where the word sample is either pre- or postfixed to the actual media.