add of arabic keys
these are the Arabic letters needed to fully type in Arabic but there is still a problem with the way kodi is handling the final output of the letter. In Arabic when the letters are used to make a work their shape would change according to the letters it get connected to. It is not like English letters where you just need to put them in sequence to type the word.

in kodi under mac osx if i use my regular hard keyboard to type, the letters would type correctly but if I try typing them on the virtual keyboard on kodi each letter is typed separate on it own which is wrong. 
this needs to be addressed by the developers. I don't know how to fix that.

thanks