configure.in: Check for taglib headers
Configure.in already asks pkg-config, but this is fooled at least on ubuntu. There is another taglib package (libtagc0-dev), that provides taglib to pkg-config, but does not provide the required headers. Those are provided by libtag1-dev.

Thus we have to check specifically for the availability of the required headers.