10-bit RGB Output
This enables the 10-bit RGB (30-bit) display mode, instead of using 8-bit RGB (24-bit).
The backbuffer is set to 10-bit. The surfaces are set to at least 10-bit.
To enable 10-bit for the display output, D3D fullscreen mode and Windows Vista or later is required.

To enable:
1) Enable fullscreen.
2) make sure "Use a fullscreen window rather than true fullscren" is not checked.
3) Create an advancedsetting.xml file like this

<advancedsettings>
    <force10BitRbgOutput>1</force10BitRbgOutput>
</advancedsettings>

This feature is not activated unless advanced settings contains the above value.