[PVR] Fix for slow metadata updates of recordings
See: http://forum.kodi.tv/showthread.php?tid=210774

metaron identified that entering the recordings screen wih 850 recordings on a Pi
was taking about 40 seconds. The database is opened and closed for every recording,
and this dominates the time taken. With the database open/close pulled out of the loop
it now takes 1.4 seconds.