[pvr] remove support for "virtual" channels
A virtual channel is/was a fake channel with a URL. These are the only
types of channels that it's been possible to add directly from the
channel manager. Alas, it doesn't actually work anymore, most likely
because no one uses this feature so no one has bothered to test it.