[confluence] fix cosmetic scrollbar shwing for half setting height
This is both cosmetic and functional.

When pulling AudioOSDsettings (ignore the fact this is videoOSDSettings as both share same xml.) the audio settings the scrollbar is present only for half a height of the last field

So really to see all settings fully you need to scroll some minimal amount for no reason.

This fixes that minor yet annoying thing.