Fix Playlist endless lopp when media fails to play, fixes #1466
When a media fails to play (ie trailers server down or any other reason), we were getting in an endless loop because HasPlayedFirstFile will never return true.
First file will never be played as it failed.
Then we create a new PL and start over adding the trailer that fails.

This leads to an endless popup asking to start from beginning / resume.