configure.in: Make SDL depend on joystick
Currently, when using --disable-sdl, the Joystick support is still enabled by default. This leads to
```
CPP     xbmc/input/ButtonTranslator.o
In file included from ButtonTranslator.cpp:44:0:
SDLJoystick.h:51:31: fatal error: SDL2/SDL_joystick.h: No such file or directory
```

This patch remove the disable-sdl switch. Instead, SDL2 is required
whenever joystick support is enabled.
If joystick support is left on auto (as it is the
default), they it is enabled when SDL2 is detected and disabled if no SDL2
is detected.