[python] Don't set PYTHONCASEOK=1.
Comment from SVN says it "should really be removed" and this fixes import
errors like:
  https://github.com/trakt/script.trakt/issues/190#issuecomment-70185992

This affected non-Windows + non-Android, but it really doesn't make any sense
considering that the vast majority of add-ons do not depend on the platform
they're running on.

tl;dr it's only causing random obscure errors.