Hide "Channels" in Guide when no channel is playing
The "Channels" view mode in the PVR Guide only has content when a tv or radio channel is playing. When the user switches through the view modes, "Channels" shows up regardless of the PVR has a channel active. It makes sense to me not letting the user navigate to an empty window, as it only confuses / makes it seem malfunctioning.

Another option would be to showing a message explaining why there's nothing to see. But I think this proposal is more elegant.