windowing/egl: add Mali support
Add the class CEGLNativeTypeMali which uses the Mali binary driver
(https://github.com/linux-sunxi/sunxi-mali) to initialize an EGL
context.

This was tested successfully on a CubieTruck and a Cubieboard.

The screen resolution is currently hard-coded to the size of my TFT
monitor.  This part will need to be rewritten.