Make Playlists home menu item appear / disappear properly, fixes 1432
We were using a dataloader property to display or not the Playlist home menu. this was not getting refreshed.

Now mediaclient jobs will trigger dataloader refresh, and dataloader will kick an event for playlists when we either have the first one, or no more.

Also fixed up a typo for `AnyOwnedServerHasPlaylists` with this commit.