Make sure we don't leave any active job, when exiting users window, part of #1450
When playing a flinged PQ from companion on locked screen, between PQ items, the Fullscreenwindow is closed, we return instantly to Users Window, and restart the full screen window again.

In That siatuation, the users fecth job is swpaned and will complete once the window has been uninitialized, causing a crash.

This patch ensures that we cancel the job if there is any, before uniniting the users window.