Don't show kb before protected user confirmation. Part of plexinc/plex-home-theater#1449 and plexinc/plex-home-theater#1450
As per comments from @chriscurtis let’s not show onscreen kb before the
user is confirmed (which would imply « i choose to login as protected
user, show me the keyboard for PIN »). Physical kb input is accepted as
soon as a protected user is selected, so both types of users should be
just as happy.