[confluence] osd save last used control & remove unsupported
Setting

<defaultcontrol always="false">100</defaultcontrol>

makes osd save the last used control and focus that next time you open
video OSD.

the issue is group list control doesnt support  defaultcontrol tag and
therefore this causes a problem when play/pause because no default
control can be set.

because of that this following is also removed.

<defaultcontrol always="true">301</defaultcontrol>

best solution is to use group control instead of group list or add
defaultcontrol support tag to group list.