[playlist] Add tags to episode smart playlists.
Allow an episode smart playlist to use the tags of the episodes tvshow
as a filter. Episodes themselves cannot be directly tagged so falling
back to the tvshow allows those tvshow tags to apply to all its
episodes.

This is handled in the same way as Genre, which is also only at the
tvshow level and not on episodes. Episode smart playlists on Genre just
use the Genres of the tvshow.