Decouple Util and StringUtil
Util included StringUtil which lead to many classes relying on it,
decouple it instead and make it explicit to include StringUtils.

This was achieved by moving the sortstringbyname from Util class
to a better fit, StringUtil. The method was also optimized to
avoid copying of the string and use the method from the StringUtil
instead.