[pi/players] Make use of new scheme to submit DTS timestamps
Latest firmware supports marking timestamps as DTS,
and it will reorder the timestamps with the decoded frame

This avoids some of the current hacks of supporting PTS and DTS
timestamps in a single field.

The current dts queue in mmal decoder gets out-of-sync for
some streams, where there isn't an exact 1:1 mapping of submitted
frame and returned picture (e.g. MBAFF or streams wih errors)
causing audio sync errors or stutter.