[pvr] check linked recordings when determining if an EPG tag is currently being recorded
Some backends may set the timer/recording start time to the actual time
the user started the recording, which means the search for a matching
timer based on the EPG tag's start and end time would fail unless the
recording was scheduled before the programme started. This would happen
regardless of whether the addon has associated an event with the
recording.