Clean up ApplicationMessenger
ApplicationMessenger had become a place to stick all kinds of convenience
methods and depending on just about everything. This PR cleans this up
and returns it back to a messaging class.

Removed the bool parameters and used naming from Win32 API with
SendMsg for blocking and PostMsg for non-blocking messaging.