Don't set FPS when the value is uknown
This commit fixes the endless loop with 100% CPU load when
setting the 0.0 FPS value.
In some configurations the RandR extension is not available
or is unable to query the Hz value.
In such cases CXRandR::GetCrtc() is not iterating in inner
loop introduced in commit cb540ac, leading to not touching
the 'hz' input variable.