[players] Remove not working SetPriority calls
When running top -H you should be able to see the priority
(niceness) of kodi threads. ActiveAESink has a niceness of -1.
(assuming OS perimissions allow that either by running as root or
through /etc/security/limits.conf)

DVDPlayerAudio should have a niceness of -1 but does not.
This is down to overloading of the SetPriority functions
we currently end up with the dummy version.

The view was that changing priority of player threads is not
useful so I've removed the calls. As it is not currently working
this should be harmless.