Fix for font corruption
CGUIFontTTFBase::DrawTextInternal iterates along the characters in the string.
For each one, it typically calls CGUIFontTTFBase::GetCharacter to (amongst
other things) locate the rasterised glyph for that character/style combination
(where style means bold/italic) within the GPU texture associated with that
font. If the texture doesn't already contain that glyph, it calls
CGUIFontTTFBase::CacheCharacter to rasterise it and add it to the texture.

The texture is made up of constant-height cells, one per glyph, and once the
texture has filled one row's worth of glyphs it wraps onto the next row; if the
texture isn't tall enough to accommodate another row then it reallocates it.

The problem is that between each CGUIFontTTFBase::GetCharacter call there is a
CGUIFontTTFBase::RenderCharacter which constructs vertices whose u/v
coordinates are the coordinates within the texture, normalised to the range
0..1. So if the texture is enlarged mid-string, then the vertices for the
earlier characters are incorrect.

I believe the reason why this wasn't detected before is that in the absence of
cacheing, the multiple re-plots of the string recalculated everything anyway,
so it was only a transient effect.