[pvr] sort equal client channel numbers by name if not sub-channels
If client channel numbers are used for ordering the channels and you have more than one backend, it's possible that you end up with different channels holding the same client channel number. In that case the channels are sorted by its sub channel number which is zero for normal channels. To have a nicely sorted list this changes extends the comparator to compare the names of the channels in case there is no sub-channel number.