[gui] introduce a dialog modality type to differ between modalities
In some cases it happens that the busy dialog is already open before the window activation kicks in.
This will results in a refuse of the activation and some pain for different addon devs because they need to hack around that issue.

This is an attempt to overcome the root cause. It adds a new dialog modality type as a replacement for the current modal flag and
specifed a new type SYSTEM_MODAL which is left-out at the check in CGUIWindowManager::ActivateWindow_Internal