[guilib] TextureDX: fix texture updates. Now there is a possibility to update texture if it was created with default usage.
By default all textures are created with usage DEFAULT. This provides the best performance. But in some cases we need to update the contents of the texture (eg teletext).
Technically, there is no possibility to upgrade the texture which was created with DEFAULT usage. So to update texture (if required) we will destroy the texture and then create a new one with DYNAMIC usage for future updates.