[videodb][infomanager] Add support for overview (plot) in movie sets.
Makes it possible to specify an overview/plot in movies sets.
Skins can retrieve it via ListItem.Plot (in fact Confluence doesn't need any
modification - "set overview" will be shown where the plot for movies is).

This changes the format of the <set> tag for movie NFO files.
Current:
  <set>My movie set</set>
Proposed:
  <set>
    <name>My movie set</name>
    <overview>This is my awesome movie set.</overview>
  </set>
The <overview> tag is optional, and the current format will still be supported
to not break existing scrapers or local NFO files.

See:
  * http://forum.kodi.tv/showthread.php?tid=237761
  * https://www.themoviedb.org/collection/121938-the-hobbit-collection
    (API does provide the "overview".)