[depends][cmake] Fix building with --with-sdk on OS X.
The environment variable was not being set, so the specified SDK was ignored by
cmake, which would instead use its own logic to determine it.

On a pre-release version of OS X (10.11), but without the Xcode beta, it
determined the appropriate SDK was MacOSX10.11 and then failed because it
couldn't find a specific header (zlib.h), because the directory didn't actually
exist.
In that case passing --with-sdk=10.10 would have fixed it, but the environment
variable wasn't set here.