Improve virtual keyboard layout for English
Previously discussed:

https://github.com/xbmc/xbmc/pull/6843

There were some objections, some request for comments from the community, but in the end nothing better got suggested other than the more simplified layout that was then added to OpenELEC:

https://github.com/OpenELEC/OpenELEC.tv/pull/4217

So far the OpenELEC change has been very positive and hasn't caused any issues, as previously feared. I'm hoping that is enough to push this through.

Other language or alternative layouts are welcome to suggest their own changes. I think it is a fair argument that their changes will be different for different areas/people/localization/whatever. In the long run, a contextually sensitive row would be better (context specific to URL entry, e-mail entry, title entry, etc), but this is the best we can do with the current code, and makes using the on-screen-keyboard much nicer (oh god, the URLs!).