[Readme]  Useful links, add bug tracker link
xexe got lost and confused and thought bug tracker wasnt in use,
due to the feature request being only in forums. DIK I gave him a gallon of coffe and an IV

xexe in forums (xe- in IRC) mentioned it and I agreed things can always
be clearer so here is nothing.