EpgContainer: use shared_ptr to manage CEpg pointers
This fixes a crash that can occur when one thread is inside
PersistAll() and the other is inside DeleteEpg().  The former does not
lock the mutex, as it operates on a std::map copy with CEpg pointers.
These CEpg instances however are deleted by the other thread.

Background information:

CEpgContainer::m_critSection protects the std::map m_epgs which
contains CEpg pointers.  All access to m_epgs must be protected with
m_critSection.  What the authors of that code didn't consider: if a
CEpg pointer escapes that protected code section, it can be
invalidated at any time.  Which implies: it is illegal to leak CEpg
pointers out of CEpgContainer::m_critSection protection!

That, however, is what commit 2ba1c43 implements, and what has been
done wrong many times before, for example GetById() and GetByChannel()
leak those pointers, explicitly.  All of that is wrong, and each one
of them can crash Kodi.

So what this commit does is ensure that all CEpg instances remain
valid as long as somebody is still holding a pointer, by wrapping all
CEpg pointers in std::shared_ptr.