When artist credits has not been populated, because only song or album tables have been queried, build a vector of artist names by splitting the artist description string.
Note this is a temporary fix as  splitting the string may not give the same artists as held in the song_artist or album_artist tables, so could lead to discrepencies. In particular in the json rpc using GetAlbums the array of artists for an album may not match the array of artistIds. This has been possible since processing of the ARTISTS tag was added. However it seems better to return artists, if sometimes they are out of alignment, than nothing.