Fix music file view generic content "songs" setting
When navigating music files we want view type to include media info so that file tag data can be displayed in file view. For this the content needs to be "songs". This original functionality was broken when deprecating the use of CGUIViewStateWindowMusicSong to make music more like video. But the initial attempt to restore that feature was too generic and caused other issues. This sets content to "songs" in a more retricted way.