[AMLCodec] Add support for libamcodec.so
The Makefiles of version 2015-01-20-4a5990f135 (latest) of the Amlogic
buildroot package will link libamcodec into "libamcodec.so" instead
of "libamplayer.so".

The buildroot package can be found at:
  http://openlinux.amlogic.com/wiki/index.php/Arm/Buildroot

The relevant file inside the archive is:
  buildroot/package/multimedia/aml_libs/src/amcodec/Makefile (line 26)

We now check if "libamcodec.so" can be loaded, and if not, we'll use
"libamplayer.so" instead for backwards compatibility with devices
running an older version.