android: fix event source evaluation
When evaluating the source of android input events we need to check the source
parameter flagwise and not absolute. The reason is that e.g. for touch events
the source parameter could be built out of multiple flags.

E.g.: source = AINPUT_SOURCE_TOUCHSCREEN | AINPUT_SOURCE_STYLUS

This ends up that xbmc is not responding to any touch events. Thus e.g. for
stylus enalbed android devices xbmc does not work without this patch.

Change-Id: I3445b9c934426a5373bcaccc5c4b0c43ee0e56d5
Signed-off-by: Danilo Krummrich <danilokrummrich@gmail.com>