Epg: fix buffer overflow in GetLastScanTime()
Passing an invalid month to CDateTime::SetDateTime() will lead to a
buffer overflow in SystemTimeToFileTime() due to these lines:

 sysTime.tm_mon = lpSystemTime->wMonth - 1;
 sysTime.tm_yday = dayoffset[sysTime.tm_mon] + (sysTime.tm_mday - 1);

This dereferences dayoffset[-1].

Instead of setting everything to zero, choose a reasonable time stamp
which was long ago in the past.

This bug should have been obvious because it was necessary to
forcefully declare the CDateTime instance as "valid".  I changed this
to be an assertion instead, to make such a bug more visible, instead
of hiding it.