VideoPlayer/VAAPI: fix use-after-free bug in Release()
The "return refCount" dereferences the CVaapiRenderPicture instance
after it was freed by vaapi->ReturnRenderPicture(this).

We know that once refCount reaches zero it will never be changed
again, but the compiler can't know that and must reload its value from
memory for the return value.  So instead of looking ip refCount again,
we simply return 0.