Remove some long-presses and some notes cleanup
This removes the arrow long-presses on the virtual keyboard, as well as creating a temp work-around for the long-press on back/backspace for Jarvis. I also removed commented notes in the keymap file that didn't apply to keyboards. They were leftover from when the keymap filewas for all input devices.