GUIWindowPVRBase: protect m_selectedItemPaths
This global variable is accessed from more than one thread, and
therefore all accesses must be serialized.

This fixes a crash (double free) bug that occurred to me when
PVRManager and EPGUpdater both attempted to destruct the same
std::string pointer that was swapped out of m_selectedItemPaths by a
move operator.