GUIWindowPVRBase: use array instead of std::map
Reduce overhead.  It is ridiculous to ues a std::map indexed by bool.

With this change, we can remove the code from the constructor that
fills the map.  This was only necessary to ensure that both keys exist
at any time, but the problem was that this map doesn't belong to a
CGUIWindowPVRBase instance; it's a global variable, and existing
values get overwritten each time a CGUIWindowPVRBase instance is
constructed.