[mmalrender] Force a SetVideoRect after a reconfigure
There has been an issue with dvd stills and a hdmi refresh rate change.
The hdmi mode change loses the currently displayed picture.
Not an issue for normal video playback as another picture will be along soon.
Not the case in DVD menus.

SetVideoRect makes the last picture redisplay and so fixes up the menu.