Limit Kodi PVR actions to PVR WIndows
Some plugins and scripts use custom controls while viewing pvr:// content. Prior code forced input dialog to popup on any numeric keypress (regardless of whether or not it was the users intentions to change the pvr channel). Limiting the dialog to PVR windows IMO is a better alternative to full screen video because it predicts the user's needs?