VideoPlayer: Avoid deadlock with m_StateSelection
Sleeping with the lock held means the OnResetDisplay callback is blocked
which blocks the application thread.

As m_StateSelection is only relased for a very few cycles it appears
that application thread struggles to obtain the lock and can stall
for 30 seconds or more.

Drop the lock and use atomics.