[DVDVideoCodec] Don't rely on initialization order for codec settings
This fixes another occurrence of the 'static initialization order
fiasco'. The two arrays mapping coding ids to settings in VAAPI and
VDPAU are statically initialized and use string constants in CSettings
that are as well statically initialized.

As both maps are only used in exactly one place, this commit moves them
to the function where they are needed (and where it's fine to access
CSettings statics). To shorten the code, they've been also converted
into a std::map.

This fixes the problem that videos couldn't be played with VDPAU/VAAPI
when using the CMake build.