Pipe crop values from the demuxer into CDemuxStream, then into CDVDStreamInfo which tags along with CDVDVideoCodec and eventually can be used to generate the CDVDVideoPicture inside CRenderManager.
Once the crop values are in the CRenderManager we use them to calculate the aspect ratio of the resulting video as well as to orient the source rectangle to crop the incoming frames.

All crop values are defaulted to zero if not found in the demuxer. This code will do nothing until paired with a version of FFmpeg which can demux the crop values and provide them in the AVStream.