cmake: Provide a way to deal with generated files in gtest_add_tests
Generated files don't exist yet, so they can't just be passed to
file(READ). Unfortunately, the GENERATED property doesn't help
that much here, since it is only visible to lists in the same
directory as the relevant source file. Since we call gtest_add_tests
at the toplevel CMakeLists.txt, it won't be visible.

Generated files which form part of the tests but don't contain
any tests themselves should be put into SUPPORT_SOURCES before
core_add_test is called. In that case, they won't be considered
as part of the library.