[cmake/osx] Adapt addon lib names and options, don't wrap addon libs
Darwin has two lib formats shared libs (dylib) and bundles (so).
Make sure to use the expected formats and extensions. We use bundles
for all libs that are not wrapped (addon libs) and dylibs for others
(libexif, libdvd).

Also use the same ARCH that is used by autotools (osx-x86 instead of
osx-x86_64-apple-darwin).