[cmake] Add additional include dir for addons
Half of the binary-addon ignore the fact, that KODI_INCLUDE_DIR
contains the kodi directory itself and include directives must not
specify it. This happens to work when building against system installed
Kodi because the compiler searches additionally in /usr/include.
When building against a different installation directory, compilation
fails with include not found errors. Instead of fixing tons of addons
with blindly copied project files, we add the root include directory as
well.