[cmake/addons] Make relative ADDON_SRC_PREFIX relative to the build directory
Currently a relative path to ADDON_SRC_PREFIX is relative to the CMake
file. This fixes the path to be relative to the current working
directory, aka. the build root directory.

This allows a consistent use of CMake from the shell where relative
paths are always understood as being relative to the cwd.