mmal_renderer: Use shared pointers to maintain the pool until the decoder has been destroyed
This fixes an issue when PVR fast channel switch changes between a SW decoded channel
(e.g. MPEG-2 without a codec licence) and a HW decoded channel (e.g. H.264).

If the pool is destroyed on the reconfigure there may still be pictures from it in the render
queue which can crash if the pool has gone when they are processed.