Reinstate the correct behaviour for 'menu' (single press)
This behaviour was broken in 779c7c4c and an attempt was made to fix it in b702e6b4. Fix didn't reinstate the original button function though, causing the menu button to change the behaviour it has had for the last decade. This change should reinstate the correct behaviour.