[cmake] Make CMake aware of textfiles that configure the build
The CMake based buildsystem is based on reading the input of textfiles
at configure-time. As CMake does not track files that are just accessed
using "file(READ)" building does not automatically retrigger a
reconfiguration when those files are changes. This commit fixes it by
either using configure_file or using the CMAKE_CONFIGURE_DEPENDS
variable.

This should remove the necessity to manually rerun CMake when files are
changed during development. (Due to globbing CMake still has to be
rerun when files are added or removed!)