[cmake] Don't run autoreconf as UPDATE_COMMAND
The ExternalProject_Add's UPDATE_COMMAND changed recently in CMake upstream to
be always executed when set. This is because it's meant to be used for updating
the sources. Kodi currently uses it however to run autoreconf and because
of the upstream change, libdvd and libcpluff are considered always
outdated and are always rebuilt.

Ref: https://github.com/Kitware/CMake/commit/bdca68388bd57f8302d3c1d83d691034b7ffa70c