[cmake] Fix GenerateVersionedFiles
This commit fixes issues with GenerateVersionedFiles:
- The generation step is called twice when a dependent file is changed.
- The generated files are always outdated when the input file is
  not modified but the timestamp is updated (touch).
- On OSX the compileinfo target fails with "Problem configuring file"
  when using the Xcode generator.

GenerateVersionedFiles internally used configure_file to generate the
output files. configure_file itself does dependency tracking which
seems to interfere with the dependencies of add_custom_command. When
an input file is changed, the script is executed and generates the
files. This however causes the other files to appear outdated as well
and the generation is triggered again.